Change potencia function to return float value

+++ lista1/exercicio12.c
@@ -6,7 +6,7 @@ int diferenca(int a, int b) { return a - b; }
 int produto(int a, int b) { return a * b; }
 int quociente(int a, int b) { return a / b; }
 int resto(int a, int b) { return a % b; }
-int potencia(int a, int b) { return (int)pow(a, b); }
+float potencia(int a, int b) { return pow(a, b); }
 
 void calcular(int a, int b) {
     printf("Operações com %d e %d:\n", a, b);
@@ -15,7 +15,7 @@ void calcular(int a, int b) {
     printf("- %d vezes %d é igual a %d;\n", a, b, produto(a, b));
     printf("- %d dividido por %d é igual a %d;\n", a, b, quociente(a, b));
     printf("- O resto da divisão de %d por %d é igual a %d;\n", a, b, resto(a, b));
-    printf("- %d elevado na %d é igual a %d;\n", a, b, potencia(a, b));
+    printf("- %d elevado na %d é igual a %.0f;\n", a, b, potencia(a, b));
 }
 
 int main() {
